refactor(app): split catalog_part5 into catalog_part8 to satisfy repo.part_unit_size
Move DLG60001 and SHEL30001 Widgetbook story hosts into a new
`catalog_part8.dart` so that `catalog_part5.dart` drops back under the
1000-physical-line ceiling enforced by `repo.part_unit_size`
(`SPEC/program/repo-lint.md`). Also wires the new part into
`catalog.dart` and leaves a breadcrumb comment in part5 pointing to the
extracted location.

No story behavior changes; all hosts, story builders, and exported
directory getters keep their original names and use cases (still wired
through `_ctWidgetbookDirectories` in `catalog.dart`).

Refs #2867
